Makefile: remove reliance on multiple venvs
Following the munet build instructions, two venvs will be
created: one manually in `~/.poetrybin`, and one by poetry
in `~/.cache/pypoetry`. `poetry install` will then install
dependencies only to the former of the two, when `make test`
expects mutest to be running in the second (since poetry
does not find VIRTUAL_ENV as being set). This results in
the failure to run any unit tests on a fresh install from
source.

This commit ensures that VIRTUAL_ENV is always set, so that
poetry understands that it is running in an existing venv
that it should also use. This consequentially prevents the
second venv from being made, and solves any unit testing
issues.

Signed-off-by: Liam Brady <lbrady@labn.net>