munet: new `*-file` config
Allows a user to configure `cmd` or `initial-cmd`
by reading from a file instead of specifying the
full command within the munet.yaml config. This
breaks up the config and should make it easier
to work with in the case of router VMs, etc.

The yang schema is also updated appropriately.

Signed-off-by: Liam Brady <lbrady@labn.net>